refactor(ui): place navbar and footer components into their respective folders
