+++ src/components/navbar/NavbarPlaceholder.vue
